# Fix an overflow bug in String.concat