@yallop 2016-09-12
One question: does OCaml actually guarantee that `int` is "modulo" — i.e. that overflow wraps?  If not, I'll need to tweak things a bit.