@alainfrisch 2016-09-12
The "Integer arithmetic" section in Pervasives' doc starts with:

> Integers are 31 bits wide (or 63 bits on 64-bit processors). All operations are taken modulo 2^31
> (or 2^63). They do not fail on overflow.