@alainfrisch 2016-09-12
(I don't know what's the maximum string length for JS backends, but I can very well imagine that it is not small enough compared to max_int to make the overflow detection safe with the unrolling hack.  That said, missing an overflow in a JS backend is probably harmless: no "segfault", just a JS exception.)