@alainfrisch 2016-09-12
Which inputs did you use for the benchmark?  The optimization is cute, but I wonder what's the impact on the whole `String.concat` (not just the length computation).  If the cost of the first `List.iter` deserves the optimization, so should the cost of the second one.