@yallop 2016-09-12
For the benchmark I used separators of length zero and one (which together account for the majority of calls to `String.concat` in the OCaml implementation), and strings of length 26.

>  If the cost of the first List.iter deserves the optimization, so should the cost of the second one.

Indeed.  And there are further opportunities to improve the performance of the function overall.  For example, since zero-length separators are common, it might be worth detecting that case upfront and avoiding n-1 calls to `unsafe_blit` that have no effect.