@dbuenzli 2016-09-12
> The unrolling uses the following observation: since Sys.max_string_length is significantly smaller than max_int

I personally don't like this assumption and Alain already pointed places where it may not hold. Since I have the same bug in Astring I have been thinking about solving the problem a bit differently. 

My idea is to check each addition of the members of the list and at the same time *count* the separators (in other words compute list length - 1). At the end of the loop we can do some arithmetic to see if 1) multiplying the sep length by the sep count would be greater than `Sys.max_string_length` 2) check if adding the total sep lengths to the accumulated lengths is greater than `Sys.max_string_length`. 

Now, of course the sep count could overflow aswell. However in order to do on a 32-bit platform your list would need to be of length > `(1 lsl 31)`. But this list would use `((1 lsl 31) * 4 * 2) / (1024 * 1024 * 1024) = 16Go` of memory which is beyond the addressable memory limit.