@dbuenzli 2016-09-12
What I said can be seen in code [here](https://github.com/dbuenzli/astring/commit/de05e8343f57007bf43f223dd164de9560fd9812).