@yallop 2016-09-12
> I don't know what's the maximum string length for JS backends

js_of_ocaml retains 32-bit OCaml's string restrictions, but has 32-bit rather than 31-bit ints:

![js_of_ocaml](https://cloud.githubusercontent.com/assets/1736496/18439534/b23c9162-78fd-11e6-885f-c63244468556.png)

Bucklescript is less constrained by convention, and has `Sys.max_string_length > max_int`.

![buckle1](https://cloud.githubusercontent.com/assets/1736496/18439737/6393d970-78fe-11e6-94a7-ad8534435719.png)
![buckle2](https://cloud.githubusercontent.com/assets/1736496/18439746/6687a9c2-78fe-11e6-86a9-16ca494bda63.png)