@bobzhang 2016-09-12
In JS vm, the string is represented by a rope, so the limit would be memory.
For this function in particular, we probably we will provide a much more efficient alternative for JS backend(shortcut `String.concat` avoid using buffer), so it may not affect us.
But it would be nice to avoid exception as a control flow if you can, local exception is preferred if  necessary.