@lefessan 2016-09-12
@yallop What are you benchmarking here exactly ? The cost of `concat`, or the cost of computing the final size ? A naive solution would be to keep the current (wrong) final size computation, but do an overflow check before the `unsafe_blit` operations ? Is it too inefficient ?