@yallop 2016-09-12
> @yallop What are you benchmarking here exactly ? The cost of concat, or the cost of computing the final size ?

The cost of computing the final size.

> A naive solution would be to keep the current (wrong) final size computation, but do an overflow check before the unsafe_blit operations ? Is it too inefficient ?

I haven't measured, but it sounds less efficient.  Wouldn't that involve two checks before each write -- for overflow of `offset+length`, and for out-of-bounds writes?  (Consider the case where the sum of the lengths of the strings is `3+2*max_int`, i.e. `1`.)