@lefessan 2016-09-14
If you are only measuring the cost of the computation of the final size, I am a bit skeptical that it is worth adding so much complexity to make it efficient, when I would assume that most of the cost of `String.concat` comes from copying the strings content into the new string.