@yallop 2016-09-14
> so much complexity 

I don't think the code is complex at all.  It's a simple recursive function that adds numbers and checks for overflow.

Which part do you consider difficult to understand?