@dbuenzli 2016-09-14
> Which part do you consider difficult to understand?

Personally I really dislike your assumption about the relationship between `Sys.max_string_length` and `max_int` at least it should be asserted in some kind of test (even though that test wouldn't be run e.g. in the js backends). 

I'm also doubtful about the real world performance impact, as @lefessan suggests the nano seconds you are fishing are likely to be lost in the noise of the actual memory copies. If that is the case there is little point to be made to go through these convolutions to implement this.