@yallop 2016-09-14
> I'm also doubtful about the real world performance impact

This was addressed in the description.

And the broader point is in danger of being lost in the noise: this PR is about fixing a bug (without degrading performance), and deliberately makes no claims about real world performance impacts.