@lefessan 2016-09-15
> Which part do you consider difficult to understand?

It's not the complexity of the code in terms of understanding (although understanding the code after 6 months takes more time than just after writing it), but in terms of number of lines. If we add 30 lines of code to fix an insignificant performance problem in each stdlib function, when just one line would be enough to fix the bug, the code of stdlib would just become unmanageable. Also, it seems easier to prove the function correct if there is only one check, instead of a call to an auxilliary function.