@bobzhang 2016-09-15
The patch is overly complicated and it might be slower in some backends (exception used as control flow here), I like @lefessan 's simple fix