@yallop 2016-09-15
@bobzhang: I have no idea what you mean by "exception used as control flow" (or what you mean by "complicated", for that matter).

Exceptions are used as, well, exceptions.