@damiendoligez 2016-09-15
My thoughts:

* This bug needs to be fixed. Why did you close this PR?
* Loop unrolling looks like a premature optimization.
* @bobzhang is talking about the cost of setting a handler for `Overflow`, although I don't see any obvious way to avoid it.
* I don't understand how any int can be larger than `max_int`. There must be something very wrong in Bucklescript to make `Sys.max_string_length > max_int` true.
* I'd like to see the one-line fix that @lefessan is talking about.