@dbuenzli 2016-09-15
> @bobzhang is talking about the cost of setting a handler for Overflow, although I don't see any obvious way to avoid it.

FWIW the [final solution](https://github.com/dbuenzli/astring/blob/1a86b39343d26b2f46ac6d673e9ac5e3f101f6c6/src/astring_string.ml#L55-L68) I have in `astring` avoids it.