@bobzhang 2016-09-15
@damiendoligez Per discussions with @dbuenzli  we will set `Sys.max_string_length` to `max_int`